requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
pandas==2.1.3
openpyxl==3.1.2
python-dotenv==1.0.0
Werkzeug==3.0.1
pytest==7.4.3
pytest-flask==1.3.0
